% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CES_marginals.R
\name{calc_CES_marginals}
\alias{calc_CES_marginals}
\title{Calculate CES Marginals}
\usage{
calc_CES_marginals(gdxName, id = "file")
}
\arguments{
\item{gdxName}{Vector of paths to \code{.gdx} files.}

\item{id}{If several \code{.gdx} files are read, an id column is appended to the
result; either \code{file}, with the paths of the originating \code{.gdx} files,
or \code{scenario}, with the content of \code{c_expname}.}
}
\value{
A \verb{data frame} with columns \code{pf} (production factor), \code{t}, \code{regi},
\code{marginal}, \code{price}, and \code{file} (path to originating \code{.gdx} file).
}
\description{
Calculate marginals on the REMIND CES function and combine them to prices.
}
\details{
Marginals are calculated analytically
\deqn{\frac{\partial V_i}{\partial V_o} = \xi_i (\theta_i \delta_i)^{\rho_o} 
      {V_o}^{1 - \rho_o} {V_i}^{\rho_o - 1}}
and prices by recursively applying the chain rule
\deqn{\pi_i = \frac{\partial V_i}{\partial V_o} \pi_o 
      \quad \forall (i,o) \in CES}
}
