% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_regionSubset_sums.R
\name{calc_regionSubset_sums}
\alias{calc_regionSubset_sums}
\title{Calculate Sums for Region Subsets}
\usage{
calc_regionSubset_sums(data, regionSubsetList)
}
\arguments{
\item{data}{A \code{\link[magclass:magclass-package]{MAgPIE}} object.}

\item{regionSubsetList}{A list of region subsets to calculate of the form
\verb{list(subset_name = c(region_A, region_B, ...)}}
}
\value{
A \code{\link[magclass:magclass-package]{MAgPIE}} object.
}
\description{
Sum up values in \code{data} for all sets of regions in \code{regionSubsetList}.
}
\examples{
calc_regionSubset_sums(population_magpie, list(xAM = c('LAM', 'NAM')))
}
\author{
Michaja Pehl
}
