% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFEprices.R
\name{calcFEprices}
\alias{calcFEprices}
\title{Read in GDX and calculate FE prices, used in convGDX2MIF.R for the
reporting.}
\usage{
calcFEprices(gdx)
}
\arguments{
\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}
}
\description{
Read in GDX and calculate FE prices, used in convGDX2MIF.R for the
reporting.
}
\examples{

\dontrun{calcFEprices(gdx)}

}
\author{
Michaja Pehl
}
