% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKayaDecomp.R
\name{calcKayaDecomp}
\alias{calcKayaDecomp}
\title{Calculates Kaya decomposition}
\usage{
calcKayaDecomp(gdx, mif = NULL, ref_year = 2005)
}
\arguments{
\item{gdx}{path to the gdx}

\item{mif}{path to a mif or data in the mif-format (if provided, the mif will be used instead of the gdx)}

\item{ref_year}{reference year for the normalization, NULL for no normalization}
}
\value{
magpie object with the kaya decomposition
}
\description{
Calculates Kaya decomposition normaized to a ref_year: CO2 FF&I/FE, FE/GDP,
GDP/Population, Population
}
\examples{

  \dontrun{
    kaya <- calcKayaDecomp(gdx,mif=NULL,ref_year=2005)
  }

}
\author{
Lavinia Baumstark
}
