% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNetTrade.R
\name{calcNetTrade}
\alias{calcNetTrade}
\title{Calculate net trade from GDX file}
\usage{
calcNetTrade(gdx, variable)
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}

\item{variable}{one of the traded commodities in REMIND: good, pecoal,
pegas, peoil, pebiolc, peur, perm}
}
\value{
Trade data as MAgPIE object
}
\description{
Calculate net trade in a commodity from a GDX file and store it into a
magpie object.
}
\examples{

  \dontrun{
    coal <- calcNetTrade(gdx, "pecoal")
    gas <- calcNetTrade(gdx, "pegas")
  }

}
\seealso{
\code{\link{readTrade}}, \code{\link{calcNetTradeValue}}
}
\author{
Niklas Roming
}
