% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNetTradeValue.R
\name{calcNetTradeValue}
\alias{calcNetTradeValue}
\title{Calculate monetary value of trade in primary energy from GDX file}
\usage{
calcNetTradeValue(gdx, variable)
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}

\item{variable}{one of the primary energy tyes in REMIND that ar: pecoal,
pegas, peoil, pebiolc, peur}
}
\value{
Value (in US$2005) of trade
}
\description{
Calculate monetary value of trade in primary energy carriers in a GDX file
and return as a MagPIE object.
}
\examples{

  \dontrun{
    coal <- calcNetTradeValue(gdx, "pecoal")
    gas <- calcNetTradeValue(gdx, "pegas")
  }

}
\seealso{
\code{\link{calcNetTrade}}, \code{\link{calcPrice}}
}
\author{
Niklas Roming
}
