% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPrice.R
\name{calcPrice}
\alias{calcPrice}
\title{Calculates normalized prices of a commodity}
\usage{
calcPrice(gdx, level = "glo", enty = "good", type = "nominal")
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}

\item{level}{spartiel resolution "glo" = global or "reg" = regional}

\item{enty}{one of the traded commodities in REMIND: good, pecoal, pegas,
peoil, pebiolc, peur, perm}

\item{type}{calculate present values (PVP(t)/PVP_Good(2005)) or nominal
values(PVP(t)/PVP_Good(t))}
}
\value{
normalized price data as MAgPIE object
}
\description{
Read PVP data of a certain commodity from a GDX file into a magpie object
and normalize it with the 2005 goods price.
}
\examples{

  \dontrun{
    coal <- calcPrice(gdx, enty = "pecoal")
    gas <- calcPrice(gdx, enty = "pegas")
  }

}
\author{
Niklas Roming
}
