% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareScenarios.R
\name{compareScenarios}
\alias{compareScenarios}
\title{Read in mif and write comparison.pdf}
\usage{
compareScenarios(mif, hist, y = c(seq(2005, 2060, 5), seq(2070, 2100,
  10)), y_hist = c(seq(1960, 2015, 1)), y_bar = c(2010, 2030, 2050,
  2100), reg = NULL, mainReg = "GLO",
  fileName = "CompareScenarios.pdf", sr15marker_RCP = NULL)
}
\arguments{
\item{mif}{a path to one or more mif-files (might be created by confGDX2MIF)}

\item{hist}{a path to one mif-file containing historical data}

\item{y}{time span for the data in line plots, default: y=c(seq(2005,2060,5),seq(2070,2100,10))}

\item{y_hist}{time span for the historical data in the line plots, default: c(seq(1960,2015,1))}

\item{y_bar}{time slides for bar plots, default: y_bar=c(2010,2030,2050,2100)}

\item{reg}{region(s) in focus, reg ="all_regi" shows all regions if the mifs contain different regions}

\item{mainReg}{region to be underlined}

\item{fileName}{name of the pdf, default = "CompareScenarios.pdf"}

\item{sr15marker_RCP}{if given, show the corresponding marker scenarios (SSP1-5) from the SR15 database in some plots.}
}
\description{
Read in all information from mif file and create
the comparison.pdf
}
\examples{

\dontrun{compareScenarios(mif_path)}

}
\author{
Lavinia Baumstark
}
