% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergenceCheck.R
\name{convergenceCheck}
\alias{convergenceCheck}
\title{Read in information form gdx and write convergenceCheck.pdf}
\usage{
convergenceCheck(
  gdx,
  y = c(seq(2005, 2060, 5), seq(2070, 2100, 10)),
  file = "convergenceCheck.pdf"
)
}
\arguments{
\item{gdx}{GDX file}

\item{y}{time span for the data in line plots, default: y=c(seq(2005,2060,5),seq(2070,2100,10))}

\item{file}{a file name the output pdf}
}
\description{
Read in all information from gdx file and create
the convergenceCheck.pdf to check tax convergence
}
\examples{

\dontrun{convergenceCheck(gdx="fulldata.gdx")}


}
\author{
Lavinia Baumstark
}
