% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convGDX2CSV_LCOE.R
\name{convGDX2CSV_LCOE}
\alias{convGDX2CSV_LCOE}
\title{Read in GDX and write LCOE .csv reporting}
\usage{
convGDX2CSV_LCOE(
  gdx,
  file = NULL,
  scen = "default",
  t = c(seq(2005, 2060, 5), seq(2070, 2110, 10), 2130, 2150)
)
}
\arguments{
\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}

\item{file}{name of the csv file which will be written}

\item{scen}{scenario name that is used in the *.mif reporting}

\item{t}{temporal resolution of the reporting, default:
t=c(seq(2005,2060,5),seq(2070,2110,10),2030,2050)}
}
\description{
Read in all information from GDX file, calculate the LCOE for pe2se technologies and create
the LCOE .csv reporting
}
\examples{
\dontrun{convGDX2CSV_LCOE(gdx,file="REMIND_LCOE_reporting.csv",scenario="default")}

}
\author{
Felix Schreyer
}
