% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convGDX2MIF_fallback_for_coupling.R
\name{convGDX2MIF_fallback_for_coupling}
\alias{convGDX2MIF_fallback_for_coupling}
\title{Read in GDX and calculate only coupling variables}
\usage{
convGDX2MIF_fallback_for_coupling(
  gdx,
  file = NULL,
  scenario = "default",
  t = c(seq(2005, 2060, 5), seq(2070, 2110, 10), 2130, 2150)
)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{file}{name of the mif file which is written, if no name is
provided a magpie object containing the reporting is returned}

\item{scenario}{scenario name that is used in the *.mif reporting}

\item{t}{temporal resolution of the reporting, default:
t=c(seq(2005,2060,5),seq(2070,2110,10),2030,2050)}
}
\description{
Read in GDX and calculate only variables that are required for MAgPIE coupling,
i.e. GHG prices and bioenergy demand.
}
\examples{

\dontrun{convGDX2MIF_fallback_for_coupling(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
David Klein
}
