% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{dashboard}
\alias{dashboard}
\title{REMIND dashboard}
\usage{
dashboard(gdx = NULL, statsFile = NULL, regionMapping = NULL,
  hist = NULL, reportfile = NULL, output_file = NULL)
}
\arguments{
\item{gdx}{GDX (fulldata.gdx) file path}

\item{statsFile}{run statistics (runstatistics.rda) file path}

\item{regionMapping}{regionMapping file name or file path (ex. "regionmappingH12.csv")}

\item{hist}{historical mif file path}

\item{reportfile}{REMIND mif report file path}

\item{output_file}{file name to save the html dashboard}
}
\description{
Create REMIND dashboard results for single runs
}
\examples{
  \dontrun{
    # loading required libraries
    library(remind)
    # creating the REMINd dashboard
    dashboard(gdx="./output/Base/fulldata.gdx",statsFile="./output/Base/runstatistics.rda",
              output_file="./output/Base/REMIND_dashboard_Base.html")
  }

}
\author{
Renato Rodrigues
}
