% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletePlus.R
\name{deletePlus}
\alias{deletePlus}
\title{Delete "+" in a *.mif reporting and save it with the addition "withoutPlus"}
\usage{
deletePlus(mif, writemif = FALSE)
}
\arguments{
\item{mif}{a path to a mif-file (might be created by confGDX2MIF) or a magpie object containing reporting}

\item{writemif}{a logical indicator whether the new mif with the adjusted names is saved under a new name containing "withoutPlus", if FALSE the new reporting is returnes as magpie object}
}
\value{
MAgPIE object - contains the reporting without "+" and "++" in the variable names
}
\description{
Delete "+" in a *.mif reporting and save it with the addition "withoutPlus"
}
\examples{

\dontrun{deletePlus(mif)}


}
\author{
Lavinia Baumstark
}
