% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimSums.R
\name{dimSums}
\alias{dimSums}
\title{dimSums}
\usage{
dimSums(x, dim = 3, na.rm = FALSE)
}
\arguments{
\item{x}{A MAgPIE-object}

\item{dim}{The dimensions(s) to sum over. A vector of dimension codes or dimension names.
See \code{\link{dimCode}} for more information}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from
the calculations?}
}
\value{
A MAgPIE object with values summed over the specified dimensions
}
\description{
Slightly modified version of the \code{dimSums} function of the magclass package.
It will expand magclass objects of length 0 by adding an element in the dimension
over which \code{dimSumsSpecial} is applied and set all values to 0 and it
will name the spatial dimension "GLO" if summation over this dimension happens.
}
\author{
Jan Philipp Dietrich
}
