% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyfunc.R
\name{dummyfunc}
\alias{dummyfunc}
\title{dummy function for the creation of the package}
\usage{
dummyfunc(n)
}
\arguments{
\item{n}{dummy n}
}
\value{
it returns n
}
\description{
dummy function for the creation of the package
}
\author{
Anastasis Giannousakis
}
