% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.copy.R
\name{gdx.copy}
\alias{gdx.copy}
\title{function for copying REMIND GDX files}
\usage{
gdx.copy(from, to)
}
\arguments{
\item{from}{path to a zipped or unzipped GDX}

\item{to}{where the unzipped GDX will go and what name it will have}
}
\description{
allows to copy a GDX file from place A to place B using the name with the
*.gz ending, regardless if it's zipped or not
}
\examples{

\dontrun{gdx.copy(path_gdx, "config/input.gdx", overwrite=TRUE)
}

}
\seealso{
\code{\link{file.copy}}
}
\author{
Anastasis Giannousakis
}
