% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunsMIFGDX.R
\name{getRunsMIFGDX}
\alias{getRunsMIFGDX}
\title{Copy fulldata.gdx and mif files from a suite of runs into one folder. This function creates the folder "./data/" in
your working directory if such folder does not exist. It will furthermore create a subfolder with the experiment name where the gdx and mif-files 
will be copied to. Care: If this subfolder is already existing, old files in this subfolder will be overwritten!}
\usage{
getRunsMIFGDX(output.folder, experiment)
}
\arguments{
\item{output.folder}{a vector with the paths to REMIND output folders of the desired runs}

\item{experiment}{name of this experiment (suite of runs)}
}
\value{
no return, function creates data folder and copies output files into it
}
\description{
Copy fulldata.gdx and mif files from a suite of runs into one folder. This function creates the folder "./data/" in
your working directory if such folder does not exist. It will furthermore create a subfolder with the experiment name where the gdx and mif-files 
will be copied to. Care: If this subfolder is already existing, old files in this subfolder will be overwritten!
}
\examples{

\dontrun{copyMIFGDX(c("Z:/Modeling/REMIND/output/BAU_2019-12-03_14.30.56/",
                       "Z:/Modeling/REMIND/output/NDC_2019-12-03_15.20.53/"),
                     experiment = "NewFeature_03_12_19")}
}
\author{
Felix Schreyer
}
