% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLCOE.R
\name{plotLCOE}
\alias{plotLCOE}
\title{Read in LCOE mif and write LCOE_plots.pdf}
\usage{
plotLCOE(mif, y = c(2015, 2020, 2030, 2040, 2050, 2060), reg = NULL,
  fileName = "LCOE_plots.pdf")
}
\arguments{
\item{mif}{a path to one or more mif-files (might be created by confGDX2MIF)}

\item{y}{time span for the data in line plots, default: y=c(2015, 2020,2030,2040,2050,2060)}

\item{reg}{region(s) in focus, reg ="all_regi" shows all regions if the mifs contain different regions}

\item{fileName}{name of the pdf, default = "LCOE_plots.pdf"}
}
\description{
Read in all information from LCOE mif file and create
the LCOE_plots.pdf
}
\author{
Felix Schreyer
}
