% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAllReportingMIFinFolder.R
\name{readAllReportingMIFinFolder}
\alias{readAllReportingMIFinFolder}
\title{Reads in all valid MIF reporting files in a folder}
\usage{
readAllReportingMIFinFolder(dir, RData = FALSE, verbose = TRUE)
}
\arguments{
\item{dir}{character. Directory containing the MIF files.}

\item{RData}{logical. If true data is saved in Rdata format to save read-in
time.}

\item{verbose}{logical. If true (un)helpful information will be display.}
}
\value{
quitte object
}
\description{
read in all valid MIF reporting files in a folder and return data as a
quitte-object. If no mif exists it searches for CSV
}
\examples{

  \dontrun{
    qd <- readAllReportingMIFinFolder('/my/magic/remind/run/')
  }

}
\author{
Anselm Schultes, Jerome Hilaire, Lavinia Baumstark, David Klein
}
