% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConsumption.R
\name{readConsumption}
\alias{readConsumption}
\title{Read Consumption from GDX file}
\usage{
readConsumption(gdx, field = "l")
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}

\item{field}{one of 'l', 'm', 's', 'lo', 'up'}
}
\description{
Read Consumption data from a GDX file into a magpie object.
}
\examples{

\dontrun{cons <- readConsumption(gdx)}

}
\author{
Jonas Hoersch
}
