% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCurrentAccount.R
\name{readCurrentAccount}
\alias{readCurrentAccount}
\title{Read Current Account from GDX file}
\usage{
readCurrentAccount(gdx)
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}
}
\description{
Read Current Account data from a GDX file into a magpie object.
}
\examples{

\dontrun{readCurrentAccount(gdx)}

}
\author{
Jonas Hoersch
}
