% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEmissions.R
\name{readEmissions}
\alias{readEmissions}
\title{Read Emissions from GDX file}
\usage{
readEmissions(gdx, emiengregi, eminegregi)
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}

\item{emiengregi}{enty that is read in from vm_emiengregi, vector is
possible e.g. c("co2","n2o"). If you do not want to add an enty from
vm_emiengregi use emiengregi=NULL}

\item{eminegregi}{enty that is read in from vm_eminegregi, vector is
possible e.g. c("co2","co2cement"). If you do not want to add an enty from
vm_emiengregi use emiengregi=NULL}
}
\description{
Read emission data from a GDX file into a magpie object. summs automatically
over all kinds of enty and vm_emiengregi and vm_emiengregi
}
\examples{

\dontrun{emi <- readEmissions(gdx,emiengregi=c("n2o","co2"),eminegregi=NULL)}

}
\author{
Lavinia Baumstark
}
