% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEnergyInvestments.R
\name{readEnergyInvestments}
\alias{readEnergyInvestments}
\title{Read Energy Investments from GDX file}
\usage{
readEnergyInvestments(gdx, field = "l")
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}

\item{field}{one of 'l', 'm', 's', 'lo', 'up')}
}
\value{
Energy Investments data as MAgPIE object
}
\description{
Read Energy Investments data from a GDX file into a MagPIE object.
}
\examples{

\dontrun{invests <- readEnergyInvestments(gdx)}

}
\author{
Jonas Hoersch
}
