% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFE.R
\name{readFE}
\alias{readFE}
\title{Read final energy from GDX file}
\usage{
readFE(gdx)
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}
}
\description{
Read final energy data from a GDX file into a magpie object.
}
\examples{

\dontrun{fe <- readFE(gdx)}

}
\author{
Lavinia Baumstark
}
