% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFuelSupplyCosts.R
\name{readFuelSupplyCosts}
\alias{readFuelSupplyCosts}
\title{Read Fuel supply costs from GDX file}
\usage{
readFuelSupplyCosts(gdx, field = "l")
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}

\item{field}{one of 'l', 'm', 's', 'lo', 'up')}
}
\value{
Fuel supply costs data as MAgPIE object
}
\description{
Read Fuel supply costs data from a GDX file into a magpie object.
}
\examples{

\dontrun{readFuelSupplyCosts(gdx)}

}
\author{
Jonas Hoersch
}
