% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGDPMER.R
\name{readGDPMER}
\alias{readGDPMER}
\title{Read GDP|MER from GDX file}
\usage{
readGDPMER(gdx, field = "l")
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}

\item{field}{one of 'l', 'm', 's', 'lo', 'up'}
}
\description{
Read GDP|MER data from a GDX file into a magpie object.
}
\examples{

\dontrun{gdp <- readGDPMER(gdx)}

}
\author{
Jonas Hoersch
}
