% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInvestmentsNonESM.R
\name{readInvestmentsNonESM}
\alias{readInvestmentsNonESM}
\title{Read Investments|Non-ESM from GDX file}
\usage{
readInvestmentsNonESM(gdx, field = "l")
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}

\item{field}{one of 'l', 'm', 's', 'lo', 'up')}
}
\value{
Investments|Non-ESM data as MAgPIE object
}
\description{
Read Investments|Non-ESM data from a GDX file into a magpie object.
}
\examples{

\dontrun{readInvestmentsNonESM(gdx)}

}
\author{
Jonas Hoersch
}
