% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNonEnergyAbatementCosts.R
\name{readNonEnergyAbatementCosts}
\alias{readNonEnergyAbatementCosts}
\title{Read non-Energy Abatement costs from GDX file}
\usage{
readNonEnergyAbatementCosts(gdx)
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}
}
\value{
non-Energy Abatement costs data as MAgPIE object
}
\description{
Read non-Energy Abatement costs data from a GDX file into a magpie object.
}
\examples{

\dontrun{readNonEnergyAbatementCosts(gdx)}

}
\author{
Jonas Hoersch
}
