% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOandMcosts.R
\name{readOandMcosts}
\alias{readOandMcosts}
\title{Read OandM costs from GDX file}
\usage{
readOandMcosts(gdx, field = "l")
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}

\item{field}{one of 'l', 'm', 's', 'lo', 'up')}
}
\value{
OandM costs data as MAgPIE object
}
\description{
Read OandM costs data from a GDX file into a magpie object.
}
\examples{

\dontrun{readOandMcosts(gdx)}

}
\author{
Jonas Hoersch
}
