% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPopulation.R
\name{readPopulation}
\alias{readPopulation}
\title{Read Population from GDX file}
\usage{
readPopulation(gdx)
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}
}
\description{
Read Population data from a GDX file into a magpie object.
}
\examples{

\dontrun{gdp <- readPopulation(gdx)}

}
\author{
Lavinia Baumstark
}
