% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPVP.R
\name{readPVP}
\alias{readPVP}
\title{Read PVP ("Present value price") from GDX file}
\usage{
readPVP(gdx, level = "glo", enty = "good")
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}

\item{level}{spartiel resolution "glo" = global or "reg" = regional}

\item{enty}{one of the traded commodities in REMIND: "good", "pecoal",
"pegas", "peoil", "pebiolc", "peur", "perm"}
}
\value{
PVP data as MAgPIE object
}
\description{
Read PVP data of a certain commodity from a GDX file into a magpie object.
}
\examples{

  \dontrun{
    coal <- readPVP(gdx, enty = "pecoal")
    gas <- readPVP(gdx, enty = "pegas")
  }

}
\author{
Jonas Hoersch
}
