% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSupplycurveBio.R
\name{readSupplycurveBio}
\alias{readSupplycurveBio}
\title{Read bioenergy supplycurve from GDX}
\usage{
readSupplycurveBio(outputdirs, userfun = function(param, x) {    
  return(param[[1]] + param[[2]] * x) }, mult_on = "all")
}
\arguments{
\item{outputdirs}{Vector providing the folders that contain the gdx files}

\item{userfun}{Function that was used to fit the supplycurve. Is needed to calculate the supplycurve correctly.
User can provide a functional form using the following syntax: \code{function(param,x)return(param[[1]] + param[[2]] * x ^param[[3]])}. This function is the default.}

\item{mult_on}{Should the multiplication factor (read from the gdx) be applied on the entire forumla ("all"), or on the slope only ("slope"). Use "all" for REMIND 2.0 (this is the default) and "slope" for REMIND 1.7}
}
\description{
Read coefficients for bioenergy prices from gdx and calculate bioenergy supplycurve. Also read the actual
demand for bioenergy from gdx.
}
\seealso{
\code{\link[remulator]{emulator}} \code{\link{calc_supplycurve}}
}
\author{
David Klein
}
