% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTimeStepWeight.R
\name{readTimeStepWeight}
\alias{readTimeStepWeight}
\title{Read Time Step Weight from GDX file}
\usage{
readTimeStepWeight(gdx)
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}
}
\value{
Time Step Weight as MAgPIE object
}
\description{
Read Time Step Weight from a GDX file into a magpie object.
}
\examples{

\dontrun{readTimeStepWeight(gdx)}

}
\author{
Jonas Hoersch
}
