% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrade.R
\name{readTrade}
\alias{readTrade}
\title{Read Trade from GDX file}
\usage{
readTrade(gdx, type, variable)
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx
file(file name is recommended as this speeds up the code)}

\item{type}{one of \code{"Imports"} or \code{"Exports"}}

\item{variable}{one of the traded commodities in REMIND: good, pecoal,
pegas, peoil, pebiolc, peur, perm}
}
\value{
Trade data as MAgPIE object
}
\description{
Read Exports or Imports of a commodity from a GDX file into a magpie object.
}
\examples{

  \dontrun{
    coal <- readTrade(gdx, "Exports", "pecoal") # Coal exports
    gas <- readTrade(gdx, "Imports", "pegas") # Gas exports
  }

}
\author{
Niklas Roming
}
