% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remind-package.R
\docType{package}
\name{remind-package}
\alias{remind-package}
\alias{remind}
\title{The REMIND R package}
\description{
Contains the REMIND-specific routines for data and model output manipulation
}
\details{
\tabular{ll}{ Package: \tab remind\cr Type: \tab Package\cr Version: \tab
7.6\cr Date: \tab 2015-04-10\cr License: \tab LGPL-3\cr LazyLoad: \tab
yes\cr }
}
\author{
Anastasis Giannousakis

Maintainer: Anastasis Giannousakis <giannou@pik-potsdam.de>
}
