% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCapitalStock.R
\name{reportCapitalStock}
\alias{reportCapitalStock}
\title{Read in GDX and calculate capital stocks, used in convGDX2MIF.R for the reporting}
\usage{
reportCapitalStock(gdx, regionSubsetList = NULL)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}
}
\value{
MAgPIE object - contains the capital stock variables
}
\description{
Read in capital stock information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportCapitalStock(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Lavinia Baumstark; Michaja Pehl
}
