% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCharts.R
\name{reportCharts}
\alias{reportCharts}
\title{Create REMIND reporting charts}
\usage{
reportCharts(
  gdx = NULL,
  regionMapping = NULL,
  hist = NULL,
  reportfile = NULL,
  chartType = "ggplot"
)
}
\arguments{
\item{gdx}{GDX file path}

\item{regionMapping}{regionMapping file name}

\item{hist}{historic file path}

\item{reportfile}{REMIND mif report file path}

\item{chartType}{plot type to include in output object. Either "plotly", "ggplot" or both c("plotly","ggplot")}
}
\description{
Create REMIND reporting plots
}
\examples{

  \dontrun{
    reportCharts(gdx="fulldata.gdx",regionMapping="./config/regionmappingH12.csv")
  }

}
\author{
Renato Rodrigues
}
