% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDiagnosis.R
\name{reportDiagnosis}
\alias{reportDiagnosis}
\title{Create REMIND Diagnosis variables and plots}
\usage{
reportDiagnosis(
  gdx = NULL,
  statsFile = NULL,
  chartType = "ggplot",
  includeData = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{statsFile}{run statistics file}

\item{chartType}{plot type to include in output object. Either "plotly", "ggplot" or both c("plotly","ggplot")}

\item{includeData}{boolean to include diagnosis data in output (default: FALSE)}
}
\description{
Create REMIND run diagnosis variables and plots
}
\examples{

  \dontrun{
    reportDiagnosis(gdx="fulldata.gdx",statsFile="runstatistics.rda")
  }

}
\author{
Renato Rodrigues
}
