% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEDGETransport.R
\name{reportEDGETransport}
\alias{reportEDGETransport}
\title{Reporting for the coupled EDGE-T Transport Sector Model (REMIND Module edge_esm)}
\usage{
reportEDGETransport(output_folder = ".", remind_root = NULL)
}
\arguments{
\item{output_folder}{path to the output folder, default is current folder.}

\item{remind_root}{path to the REMIND root directory, defaults to two levels up from output_folder.}
}
\description{
Data is loaded from the EDGE-T subfolder in the output folder.
The input files can be (re-) generated calling
`Rscript EDGETransport.R --reporting`
from the output folder.
}
\details{
*Warning* The function modifies the "REMIND_generic_<scenario>.mif" file by appending the
additional reporting variables and replaces the "_withoutPlus" version.

Region subsets are obtained from fulldata.gdx
}
\author{
Alois Dirnaichner Marianna Rottoli
}
