% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEmi.R
\name{reportEmi}
\alias{reportEmi}
\title{Read in GDX and calculate emissions, used in convGDX2MIF.R for the
reporting}
\usage{
reportEmi(gdx, output = NULL, regionSubsetList = NULL)
}
\arguments{
\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}
}
\description{
Read in emission information from GDX file, information used in
convGDX2MIF.R for the reporting
}
\examples{

  \dontrun{reportEmi(gdx)}
}
\author{
Jessica Streffler, Lavinia Baumstark, Michaja Pehl
}
