% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEmiAirPol.R
\name{reportEmiAirPol}
\alias{reportEmiAirPol}
\title{Read in GDX and calculate air pollution emissions, used in convGDX2MIF.R for 
the reporting}
\usage{
reportEmiAirPol(gdx, regionSubsetList = NULL)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}
}
\value{
MAgPIE object - contains the emission variables
}
\description{
Read in air pollution emission information from GDX file, information used in 
convGDX2MIF.R for the reporting
}
\examples{
\dontrun{reportEmiAirPol(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Antoine Levesque, Jerome Hilaire
}
