% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEmployment.R
\name{reportEmployment}
\alias{reportEmployment}
\title{Computes the employment values (jobs) across different sectors}
\usage{
reportEmployment(gdx)
}
\arguments{
\item{gdx}{A gdx object}
}
\value{
A magpie object
}
\description{
This function returns a magpie object containing the reporting the jobs for different technologies
}
\examples{

  \dontrun{
    reportEmployment(gdx)
  }
}
\author{
Aman Malik
}
