% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEnergyInvestment.R
\name{reportEnergyInvestment}
\alias{reportEnergyInvestment}
\title{Read in GDX and calculate prices, used in convGDX2MIF.R for the reporting}
\usage{
reportEnergyInvestment(gdx, regionSubsetList = NULL)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}
}
\value{
MAgPIE object - contains the price variables
}
\description{
Read in price information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportEnergyInvestment(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Anastasis Giannousaki
}
