% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPrices.R
\name{reportPrices}
\alias{reportPrices}
\title{Read in GDX and calculate prices, used in convGDX2MIF.R for the reporting}
\usage{
reportPrices(gdx, output = NULL, regionSubsetList = NULL)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}
}
\value{
MAgPIE object - contains the price variables
}
\description{
Read in price information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportPrices(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
David Klein
}
