% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTax.R
\name{reportTax}
\alias{reportTax}
\title{Read in GDX and calculate tax, used in convGDX2MIF.R for the reporting}
\usage{
reportTax(gdx, regionSubsetList = NULL)
}
\arguments{
\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}
}
\description{
Read in tax information from GDX file, information used in convGDX2MIF.R for
the reporting
}
\examples{

\dontrun{reportTax(gdx)}

}
\author{
Lavinia Baumstark, Christoph Bertram
}
