% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTechnology.R
\name{reportTechnology}
\alias{reportTechnology}
\title{Read in GDX and calculate technology information, used in convGDX2MIF.R for the reporting}
\usage{
reportTechnology(gdx, output = NULL, regionSubsetList = NULL)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}
}
\value{
MAgPIE object - contains the technology variables
}
\description{
Read in technology information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportTechnology(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Michaja Pehl, Lavinia Baumstark
}
