% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTrade.R
\name{reportTrade}
\alias{reportTrade}
\title{Read in GDX and calculate trade, used in convGDX2MIF.R for the reporting}
\usage{
reportTrade(gdx, regionSubsetList = NULL)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}
}
\value{
MAgPIE object - contains the price variables
}
\description{
Read in trade information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportTrade(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Lavinia Baumstark, Christoph Bertram, Anselm Schultes
}
