% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolRegionSubsets.R
\name{toolRegionSubsets}
\alias{toolRegionSubsets}
\title{toolRegionSubsets
Returns a list of parent regions that are equal to a child mapping union of region mappings.}
\usage{
toolRegionSubsets(
  gdx = NULL,
  map = NULL,
  parentMapping = NULL,
  singleMatches = FALSE,
  removeDuplicates = TRUE,
  regionIndex = NULL,
  countryIndex = NULL
)
}
\arguments{
\item{gdx}{a GDX as created by readGDX, or the file name of a gdx.
Gdx file containing child mapping regions list.}

\item{map}{alternative to gdx file. You can also provide a child mapping csv file.}

\item{parentMapping}{parent mapping or csv file (if NULL it uses the "regionmappingH12.csv").}

\item{singleMatches}{if true, includes single element matches in the return list.
Default: FALSE}

\item{removeDuplicates}{if true, removes from the return list identic matches (a region with
the same name and countries for both parent and child mappings). Default: TRUE}

\item{regionIndex}{Name or index of the region column to be used in the parent mapping (if not
set the column with less unique elements from the last two columns will be used).}

\item{countryIndex}{Name or index of the country column to be used in  parent mapping (if not
set the column with more unique elements from the last two columns will be used).}
}
\value{
return: returns a list of parent regions that are equal to a child mapping union
of region mappings.
}
\description{
toolRegionSubsets
Returns a list of parent regions that are equal to a child mapping union of region mappings.
}
\examples{

\dontrun{toolRegionSubsets(gdx)}

}
\author{
Renato Rodrigues
}
