% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationREMIND.R
\name{validationREMIND}
\alias{validationREMIND}
\title{validation}
\usage{
validationREMIND(gdx, hist, file = "validation.pdf", reportfile = NULL, ...)
}
\arguments{
\item{gdx}{GDX file}

\item{hist}{Validation data.All formats allowed which can be converted to quitte (including characters containing the path to a mif file)}

\item{file}{a file name the output pdf}

\item{reportfile}{file name to which a backup of the magpie reporting should be written (file ending should be ".mif"). No report written if set to NULL.}

\item{...}{additional arguments supplied to the validationpdf function}
}
\description{
Create Validation pdf from REMIND output and corresponding historical.mif
}
\examples{

  \dontrun{
    validation(gdx="fulldata.gdx",hist="historical.mif")
  }

}
\author{
Lavinia Baumstark
}
