% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationSummary.R
\name{validationSummary}
\alias{validationSummary}
\title{validationSummary}
\usage{
validationSummary(
  gdx,
  hist,
  reportfile = NULL,
  outfile = "validationSummary.pdf",
  ...
)
}
\arguments{
\item{gdx}{GDX file}

\item{hist}{Validation data.All formats allowed which can be converted to quitte (including characters containing the path to a mif file)}

\item{reportfile}{file name to which a backup of the magpie reporting should be written (file ending should be ".mif"). No report written if set to NULL.}

\item{outfile}{file name of the output pdf}

\item{...}{additional arguments supplied to the validationpdf function}
}
\description{
Create validationSummary.pdf from REMIND output and corresponding historical.mif with most important variables
}
\examples{

  \dontrun{
    validationSummary(gdx="fulldata.gdx",hist="historical.mif")
  }

}
\author{
David Klein, Lavinia Baumstark
}
