% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLCOE.R
\name{plotLCOE}
\alias{plotLCOE}
\title{Read in LCOE mif and write LCOE_plots.pdf}
\usage{
plotLCOE(
  LCOEfile,
  gdx,
  y = c(2015, 2020, 2030, 2040, 2050, 2060),
  reg = "all_regi",
  fileName = "LCOE_plots.pdf"
)
}
\arguments{
\item{LCOEfile}{a path to the LCOE reporting csv file of the scenario to be plotted}

\item{gdx}{gdx file of run}

\item{y}{time span for the data in line plots, default: y=c(2015, 2020,2030,2040,2050,2060)}

\item{reg}{region(s) in focus, reg ="all_regi" shows all regions if the mifs contain different regions}

\item{fileName}{name of the pdf, default = "LCOE_plots.pdf"}
}
\description{
Read in all information from LCOE mif file and create
the LCOE_plots.pdf
}
\author{
Felix Schreyer
}
