% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEmployment.R
\name{reportEmployment}
\alias{reportEmployment}
\title{Computes the employment values (jobs) across different sectors}
\usage{
reportEmployment(gdx, improvements)
}
\arguments{
\item{gdx}{A gdx file output from a REMIND run}

\item{improvements}{source of employment factor data, options "None", "Dias","CEEW", "Dias+CEEW"}
}
\value{
A magpie object
}
\description{
This function returns a magpie object containing the reporting the jobs for different technologies
}
\examples{

  \dontrun{
    reportEmployment(gdx,improvements="Dias")
  }
}
\author{
Aman Malik
}
