% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLCOE.R
\name{reportLCOE}
\alias{reportLCOE}
\title{Read in GDX and LCOE, used in convGDX2MIF.R for the reporting}
\usage{
reportLCOE(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains LCOE cost parts per pe2se technology
}
\description{
Read in LCOE information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportLCOE(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Robert Pietzcker, Lavinia Baumstark, Felix Schreyer
}
