% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMOFEX.R
\name{reportMOFEX}
\alias{reportMOFEX}
\title{Calculate and report relevant fossil fuel cost and quantity variables from MOFEX standalone model.}
\usage{
reportMOFEX(gdx, gdx_ref = NULL, file = NULL, scenario = "default")
}
\arguments{
\item{gdx}{the filepath to a gdx}

\item{gdx_ref}{reference-gdx for policy costs, a GDX as created by readGDX, or the file name of a gdx}

\item{file}{path to output .mif file}

\item{scenario}{scenario name that is used in the *.mif reporting}
}
\value{
MAgPIE object - contains the output from a MOFEX standalone run
}
\description{
Calculate and report relevant fossil fuel cost and quantity variables from MOFEX standalone model.
}
\examples{

\dontrun{reportMOFEX(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}} \code{\link{reportPE}} \code{\link{reportCosts}}
}
\author{
Steve Bi
}
